Text.Regex.XMLSchema.Generic.StringLike

Description

Documentation

class (Eq a, IsString a, Show a) => StringLike a where

WARNING: This StringLike class is not intended for use outside this regex library. It provides an abstraction for String's as used inside this library. It allows the library to work with String (list of Char), ByteString.Char8, ByteString.Lazy.Char8, Data.Text and Data.Text.Lazy.

The class is similar to the StringLike class in the tagsoup package

Minimal complete definition

emptyS, uncons, takeS, dropS, appendS, toString

Methods

emptyS :: a

uncons :: a -> Maybe (Char, a)

nullS :: a -> Bool

headS :: a -> Char

takeS :: Int -> a -> a

dropS :: Int -> a -> a

appendS :: a -> a -> a

concatS :: [a] -> a

toString :: a -> String